requirements.txt
typer==0.12.3
pydantic==2.8.2
requests==2.32.3
beautifulsoup4==4.12.3
tqdm==4.66.4
pdfminer.six==20240706
